PythonGUI/requirements.txt
numpy==1.24.2
PyQt5==5.15.9
PyQtChart==5.15.6
pyserial==3.5
python-can==4.1.0
cantools==38.0.2